tests/e2e/requirements.txt
# E2E Testing Requirements
pytest==7.4.3
Appium-Python-Client==3.1.0
selenium==4.15.2
requests==2.31.0
